--- cartridges/int_gtm/cartridge/templates/default/common/layout/checkout.isml
@@ -1,53 +1,0 @@
-<iscontent type="text/html" charset="UTF-8" compact="true"/>
-
-<isinclude template="/components/modules" sf-toolkit="off" />
-
-<!DOCTYPE html>
-<html lang="${require('dw/util/Locale').getLocale(request.getLocale()).getLanguage()}">
-    <head>
-        <!--[if gt IE 9]><!-->
-            <isinclude sf-toolkit="off" template="/common/scripts" />
-        <!--<![endif]-->
-        <isinclude template="/common/htmlHead" />
-
-        <isinclude url="${URLUtils.url('GtmIdSet-SetHead')}"/>
-
-        <isif condition="${pdict.canonicalUrl}" >
-            <link rel="canonical" href="${pdict.canonicalUrl}"/>
-        </isif>
-        <isactivedatahead/>      
-        <isinclude template="/components/schema" />
-
-    </head>
-    <body>
-        ${dw.system.HookMgr.callHook('app.template.beforeHeader', 'beforeHeader', pdict) || ''}
-
-        <div class="page">
-            <isinclude template="/components/header/pageHeaderNomenu" />
-            <div role="main" id="maincontent">
-                <isreplace/>
-            </div>
-            <isinclude template="/components/footer/pageFooter" />
-        </div>
-
-        <!--[if lt IE 10]>
-            <isinclude sf-toolkit="off" template="/common/scripts" />
-        <![endif]-->
-
-        <div class="error-messaging"></div>
-        <div class="modal-background"></div>
-        <iscontentasset aid="cookie_hint" />        
-      
-        <iscomment>
-            hook for Marketing Cloud connector & other integration which need to inject
-            logic at the page end
-            IMPORTANT: Note that this hook will be called to cached as well as uncached pages
-                        which means you need to put privacy information into another remote include
-        </iscomment>
-
-        ${dw.system.HookMgr.callHook('app.template.afterFooter', 'afterFooter', pdict) || ''}
-
-        <isinclude url="${URLUtils.url('ConsentTracking-Check')}"/>
-        <isinclude url="${URLUtils.url('GtmIdSet-SetBody')}"/>
-    </body>
-</html>
