+++ cartridges/int_gtm/cartridge/templates/default/common/layout/page.isml
@@ -8,47 +8,40 @@
 
 <!DOCTYPE html>
 <html lang="${require('dw/util/Locale').getLocale(request.getLocale()).getLanguage()}">
+    <head>
+        <!--[if gt IE 9]><!-->
+            <isinclude sf-toolkit="off" template="/common/scripts" />
+        <!--<![endif]-->
+        <isinclude template="/common/htmlHead" />
+        <isif condition="${pdict.canonicalUrl}" >
+            <link rel="canonical" href="${pdict.canonicalUrl}"/>
+        </isif>
+        <isactivedatahead/>
+        <isinclude template="/components/schema" />
+    </head>
+    <body>
+        ${dw.system.HookMgr.callHook('app.template.beforeHeader', 'beforeHeader', pdict) || ''}
 
-<head>
-    <isinclude url="${URLUtils.url('GtmIdSet-SetHead')}"/>
-
-    <!--[if gt IE 9]><!-->
-    <isinclude sf-toolkit="off" template="/common/scripts" />
-    <!--<![endif]-->
-    <isinclude template="/common/htmlHead" />
-    <isif condition="${pdict.canonicalUrl}">
-        <link rel="canonical" href="${pdict.canonicalUrl}" />
-    </isif>
-    <isactivedatahead/>
-    <isinclude template="/components/schema" />
-</head>
-
-<body>
-    <isinclude url="${URLUtils.url('GtmIdSet-SetBody')}"/>
-    
-    ${dw.system.HookMgr.callHook('app.template.beforeHeader', 'beforeHeader', pdict) || ''}
-
-    <div class="page" data-action="${pdict.action}" data-querystring="${pdict.queryString}">
-        <isinclude template="/components/header/pageHeader" />
-        <div role="main" id="maincontent">
-            <isreplace/>
+        <div class="page" data-action="${pdict.action}" data-querystring="${pdict.queryString}" >
+            <isinclude template="/components/header/pageHeader" />
+            <div role="main" id="maincontent">
+                <isreplace/>
+            </div>
+            <isinclude template="/components/footer/pageFooter" />
         </div>
-        <isinclude template="/components/footer/pageFooter" />
-    </div>
-    <div class="error-messaging"></div>
-    <div class="modal-background"></div>
-    <iscontentasset aid="cookie_hint" />
-    <!--[if lt IE 10]>
+        <div class="error-messaging"></div>
+        <div class="modal-background"></div>
+        <iscontentasset aid="cookie_hint" />
+         <!--[if lt IE 10]>
             <isinclude sf-toolkit="off" template="/common/scripts" />
         <![endif]-->
-    <iscomment>
-        hook for Marketing Cloud connector & other integration which need to inject
-        logic at the page end
-        IMPORTANT: Note that this hook will be called to cached as well as uncached pages
-        which means you need to put privacy information into another remote include
-    </iscomment>
-    ${dw.system.HookMgr.callHook('app.template.afterFooter', 'afterFooter', pdict) || ''}
-    <isinclude url="${URLUtils.url('ConsentTracking-Check')}" />
-
-</body>
+        <iscomment>
+            hook for Marketing Cloud connector & other integration which need to inject
+            logic at the page end
+            IMPORTANT: Note that this hook will be called to cached as well as uncached pages
+                        which means you need to put privacy information into another remote include
+        </iscomment>
+        ${dw.system.HookMgr.callHook('app.template.afterFooter', 'afterFooter', pdict) || ''}
+        <isinclude url="${URLUtils.url('ConsentTracking-Check')}"/>
+    </body>
 </html>
