+++ cartridges/int_gtm/hooks.json
@@ -1,0 +1,12 @@
+{
+    "hooks": [
+        {
+            "name": "app.template.htmlHead",
+            "script": "./cartridge/scripts/hooks/gtm/gtmhooks.js"
+        },
+        {
+            "name": "app.template.beforeHeader",
+            "script": "./cartridge/scripts/hooks/gtm/gtmhooks.js"
+        }
+    ]
+}
