+++ cartridges/int_gtm/package.json
@@ -1,0 +1,3 @@
+{
+    "hooks": "./hooks.json"
+}
